+++ docs/ReleaseNotes.md
@@ -3,6 +3,7 @@ _0.2.0_
 - Remove InitializeSchema API and replace with Initialize.
 - Fix for exception on unknown format, now defaults to .NET default double format.
 - Fix for schema mappings falling back to ordinal not working.
+- Fix for empty shared string values created by AG Grid.
 
 _0.1.12_
 - Fix for reading .xlsx files with inline-string values, believed to be created with Apache POI.